create_poll: don't send emails in case of DB transaction errors
This commit adds an error message in the STEP 3 of the poll creation
form in case there is a DB transactions error.

It also makes sure to not send emails in case of those DB transaction
errors.

+++ app/classes/Framadate/Services/PollService.php
@@ -206,16 +206,24 @@ class PollService {
         $admin_poll_id = $ids[1];
 
         // Send confirmation by mail if enabled
-        // Everything went well
-        $this->notificationService->sendPollCreationMails($form->admin_mail, $form->admin_name, $form->title, $poll_id, $admin_poll_id);
+        if (!is_null($ids)) {
+            // Everything went well
+            $this->notificationService->sendPollCreationMails($form->admin_mail, $form->admin_name, $form->title, $poll_id, $admin_poll_id);
 
-        // Clean Form data in session
-        $this->sessionService->removeAll('form');
-        // Creation message
-        $this->sessionService->set("Framadate", "messagePollCreated", TRUE);
-
-        // Delete old polls
-        $this->purgeService->repeatedCleanings();
+            // Clean Form data in session
+            $this->sessionService->removeAll('form');
+            // Creation message
+            $this->sessionService->set("Framadate", "messagePollCreated", TRUE);
+            // Delete old polls
+            $this->purgeService->repeatedCleanings();
+        } else {
+            // Add an error in the form
+            $form->errors = [
+                __('Error', 'GenericErrorPollCreation')
+            ];
+            // TODO: change this to use sessionService function
+            $_SESSION['form'] = serialize($form);
+        }
 
         return $admin_poll_id;
     }
